refactor: move domain objects into Domain folder
