+++ jobs/Backend/Task/src/Domain/Currency.cs
@@ -1,4 +1,4 @@
-﻿namespace ExchangeRateUpdater
+﻿namespace ExchangeRateUpdater.Domain
 {
     public class Currency
     {
