+++ jobs/Backend/Task/src/Domain/ExchangeRate.cs
@@ -1,4 +1,4 @@
-﻿namespace ExchangeRateUpdater
+﻿namespace ExchangeRateUpdater.Domain
 {
     public class ExchangeRate
     {
